+++ src/ToggleSwitch.ts
@@ -1,0 +1,18 @@
+/**
+ * ToggleSwitch component entry point
+ *
+ * Peer dependencies required to use this component:
+ * - `@emotion/react`
+ * - `react`
+ * - `react-dom`
+ * - `typescript`
+ * - `react-aria-components`
+ *
+ * If you only need the built CSS (./component/toggleSwitch.css),
+ * you don't need to install these.
+ */
+export { ToggleSwitch } from './components/ToggleSwitch/ToggleSwitch';
+export type { ToggleSwitchProps } from './components/ToggleSwitch/types';
+export type { PartialToggleSwitchTheme as ToggleSwitchTheme } from './components/ToggleSwitch/styles';
+export { componentToggleSwitch } from './styleD/build/typescript/component/toggleSwitch';
+export type { ComponentToggleSwitch } from './styleD/build/typescript/component/toggleSwitch';