+++ CHANGELOG.md
@@ -1,5 +1,75 @@
 # @guardian/stand
 
+## 0.0.73
+
+### Patch Changes
+
+- e983622: Correcting exports for a number of components, non breaking
+- 59572c4: Add pagination component
+
+## 0.0.72
+
+### Patch Changes
+
+- 7a9fbfc: Add `ToggleSwitch` component
+- 991eb6b: BREAKING: Update semantic tokens, naming changed:
+
+  - `{semantic.colors.fill.selected}` -> `{semantic.colors.fill.selectedStrong}`
+    - Color has also been updated from `{base.colors.cyan.300}` to `{base.colors.cyan.400}`
+
+  Added a number of semantic colour tokens.
+  Added `{semantic.sizing.height.xxxxs}`.
+
+## 0.0.71
+
+### Patch Changes
+
+- 8fb8e10: Add the accessible, responsive Table component with React Aria semantics, composable CSS Grid layouts, design tokens, and rounded corners.
+
+## 0.0.70
+
+### Patch Changes
+
+- 8ea936c: Add `TextListInput` component
+
+## 0.0.69
+
+### Patch Changes
+
+- 4a59ead: Support TypeScript 6
+
+## 0.0.68
+
+### Patch Changes
+
+- dcaa35a: Fix - `Typography` - Allow `element` prop to be any string, also inherit autocomplete list from `JSX.IntrinsicElements`
+
+## 0.0.67
+
+### Patch Changes
+
+- 0a00eb2: Add React Aria 1.20.0 to peer dependency range
+- a0082b4: Add `SidebarStepperNavigation` for presenting and navigating multi-step workflows
+
+## 0.0.66
+
+### Patch Changes
+
+- 6f24c77: Adds the Badge component to Stand
+
+## 0.0.65
+
+### Patch Changes
+
+- 502e9d6: Add `size` prop and small variant to `Select` component
+
+## 0.0.64
+
+### Patch Changes
+
+- 2d1a2ea: Add `UserFeedbackSummary` component
+- 622180a: Updates to design tokens issued in 22 July 2026
+
 ## 0.0.63
 
 ### Patch Changes